+++ modules/Shmuelie.Windows/Public/PlatformGuards.ps1
@@ -17,3 +17,9 @@ function Assert-WindowsOnly {
         throw "$CommandName is only supported on Windows."
     }
 }
+function Test-IsElevated {
+    [CmdletBinding()]
+    param()
+
+    return ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
+}